(MODULES-2299) Refactor vendors_dsc?
`vendors_dsc?` should not include the question mark in the name. That
can confuse Puppet when it creates the predicate
`Puppet.features.featurename?` where in this case it may end up
creating `Puppet.features.vendors_dsc??`. Rename the feature to
`vendors_dsc` instead.

Additionally, having the symlink module code inside the feature is
confusing Linux Masters, with the failure happening on the early return
false at
https://github.com/puppetlabs/puppetlabs-dsc/blob/662a2961cd4f20e2e11fb82ae4778ca53071e9e5/lib/puppet/feature/vendors_dsc.rb#L7
Instead, move the symlink code out to it's own module and require that
in the feature add.

Paired with James Pogran <james pogran at puppetlabs>
Paired with Josh Cooper <josh at puppetlabs>
